Enhancement: Move tests into ObjectProphecy namespace
